backend/requirements.txt
flask
flask-cors
yfinance
supabase
python-dotenv
flask-cors
